requirements.txt
streamlit>=1.28.0
google-genai>=0.2.0
pydantic>=2.0.0
